+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [2.0.0] - 5 Januari 2023
+- Update stepper to allow for growable content in rows
+- Refactor structure of app for better maintenance
+- Rename showAllStepExpandedOne to showOnlyCurrentStep
+
 ## [1.0.2] - 25 November 2022
 
 - Updated the basic styling of the stepper component